refactor: 엔티티 패키지 수정 및 검증로직 개선
