+++ src/test/java/roomescape/service/query/ThemeQueryServiceTest.java
@@ -12,7 +12,7 @@ import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
-import roomescape.domain.Theme;
+import roomescape.domain.reservation.slot.Theme;
 import roomescape.dto.theme.ThemeResponseDto;
 import roomescape.repository.JpaThemeRepository;
